% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{all_possible}
\alias{all_possible}
\title{Helper for getting all data on specified variables with N observed}
\usage{
all_possible(model, N, vars = NULL, condition = TRUE)
}
\description{
Helper for getting all data on specified variables with N observed
}
\examples{
model <- make_model("X->M->Y")
gbiqqtools:::all_possible(model, N=2, vars = c("X", "M"))
gbiqqtools:::all_possible(model, N=2, vars = c("X", "Y"), condition = "Y==0")
}
