% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_event_data}
\alias{check_event_data}
\title{Helper: order event data and add strategy family}
\usage{
check_event_data(df, model)
}
\arguments{
\item{df}{dataframe with event data}

\item{model}{A \code{causal_model}. A model object generated by \code{\link[gbiqq]{make_model}}}
}
\description{
Helper: order event data and add strategy family
}
\examples{
model <- make_model("X -> M -> Y")
df <- data.frame(M = c(1, NA), X = c(1,1), Y = c(NA,NA)) \%>\%
 collapse_data(model) \%>\%
 filter(count == 1)
df
gbiqqtools:::check_event_data(df, model)

}
