% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_inherit_arguments.R
\name{CQtools_internal_inherit_params}
\alias{CQtools_internal_inherit_params}
\title{Create argument documentation to inherit}
\usage{
CQtools_internal_inherit_params(
  model,
  reference_model,
  analysis_model,
  parameters,
  given,
  sims,
  iter,
  chains,
  query,
  strategy,
  observed
)
}
\arguments{
\item{model}{A \code{causal_model}. A model object generated by \code{\link[CausalQueries]{make_model}}}

\item{reference_model}{A causal model as created by \code{\link[CausalQueries]{make_model}}}

\item{analysis_model}{A causal model as created by \code{\link[CausalQueries]{make_model}}}

\item{parameters}{A vector of real numbers in [0,1]. Values of parameters to specify (optional). By default, parameters is drawn from \code{model$parameters_df}}

\item{given}{A conditioning set as a character string that evaluates to a logical, for example 'Y==1'}

\item{sims}{Integer, number of estimand draws, defaults to max(sims, 4000)}

\item{iter}{Integer, passed to stan, defaults to 4000}

\item{chains}{Integer, passed to stan, defaults to 4}

\item{query}{A character string. An expression defining nodal types to interrogate \code{\link[CausalQueries]{reveal_outcomes}}}

\item{strategy}{A set of node to be sought}

\item{observed}{A data.frame with observations.}
}
\description{
Create argument documentation to inherit
}
\keyword{internal}
