% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expected_learning.R
\name{expected_learning}
\alias{expected_learning}
\title{Expected learning}
\usage{
expected_learning(
  model,
  query,
  strategy = NULL,
  given = NULL,
  parameters = NULL
)
}
\arguments{
\item{model}{A \code{causal_model}. A model object generated by \code{\link[CausalQueries]{make_model}}}

\item{query}{A character string. An expression defining nodal types to interrogate \code{\link[CausalQueries]{reveal_outcomes}}}

\item{strategy}{A set of node to be sought}

\item{given}{A conditioning set as a character string that evaluates to a logical, for example 'Y==1'}

\item{parameters}{A vector of real numbers in [0,1]. Values of parameters to specify (optional). By default, parameters is drawn from \code{model$parameters_df}}
}
\description{
Expected reduction in variance from one step data collection strategy
}
\examples{
# Reduction in variance given monotonic X -> M1 -> M2 -> Y model

model <- make_model("X -> M1 -> M2 -> Y") \%>\%
  set_restrictions(labels = list(M1 = "10", M2 = "10", Y = "10"))
el <- expected_learning(model, query = "Y[X=1]>Y[X=0]",
                  strategy = c("X", "M2"), given = "Y==1")
attr(el, "results_table")

el2 <- expected_learning(model, query = "Y[X=1]>Y[X=0]",
                  strategy = c("M1"),
                  given = "Y==1 & X==1 & M2==1")
attr(el2, "results_table")

# No strategy
expected_learning(model, query = "Y[X=1]>Y[X=0]")

# No givens
expected_learning(model, query = "Y[X=1]>Y[X=0]",
strategy = c("M1"))
expected_learning(model, query = "Y[X=1]>Y[X=0]",
strategy = c("M1"), given = "Y==1")

model <-  make_model("S -> C -> Y <- R <- X; X -> C -> R") \%>\%
set_restrictions(labels =  list(C = "1110", R = "0001", Y = "0001"),  keep = TRUE)

expected_learning(model,
query = list(COE = "(Y[S=0] > Y[S=1])"),
strategy = "C", given = "Y==1 & S==0")

expected_learning(model,
query = list(COE = "(Y[X=1] > Y[X=0])"),
strategy = "S", given = "X==0 & Y==0")
}
