% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbiqqtools.R
\docType{package}
\name{gbiqqtools}
\alias{gbiqqtools}
\alias{gbiqqtools-package}
\title{gbiqqtools: A package for choosing research strategies.}
\description{
gbiqqtools: A package for choosing research strategies.
}
