% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_probs.R
\name{get_data_probs}
\alias{get_data_probs}
\title{Get data probabilities}
\usage{
get_data_probs(model, data, parameters = NULL)
}
\arguments{
\item{model}{A \code{causal_model}. A model object generated by \code{\link[CausalQueries]{make_model}}}

\item{data}{Data in long format}

\item{parameters}{A vector of real numbers in [0,1]. Values of parameters to specify (optional). By default, parameters is drawn from \code{model$parameters_df}}
}
\description{
Takes in a matrix of possible (single case) observations and returns the probability of each.
FLAG: Some redundancy with make_data_probabilities
}
\examples{
model <- make_model("X->Y")
data <- simulate_data(model, n = 4)
get_data_probs(model, data)
}
