% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_likelihoods.R
\name{get_loo_likelihood}
\alias{get_loo_likelihood}
\title{Get leave one out data likelihood}
\usage{
get_loo_likelihood(data, model, sims = 100)
}
\arguments{
\item{data}{Data in long format}

\item{model}{A \code{causal_model}. A model object generated by \code{\link[CausalQueries]{make_model}}}

\item{sims}{Number of simulations}
}
\description{
Assesses data likelihood using leave out procedure
}
\examples{
model <- make_model("X->Y")
data <- simulate_data(model, n = 4)
get_loo_likelihood(model, data)
}
