% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_nodal_joint_probability}
\alias{get_nodal_joint_probability}
\title{Get joint distribution of nodal types}
\usage{
get_nodal_joint_probability(
  model,
  parameters = NULL,
  generic_parameters = TRUE
)
}
\arguments{
\item{generic_parameters}{Logical. Whether to require selection of a generic parameter. Defaults to TRUE.}
}
\description{
Identifies possible conditional probabilities of nodal types. May be used to identify patterns of non independence.
}
\examples{
model <- make_model('X -> Y') \%>\%
  set_confound(list('X <-> Y'))

get_nodal_joint_probability(model)

}
\keyword{internal}
