% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data_probabilities.R
\name{make_data_probabilities}
\alias{make_data_probabilities}
\title{Generates a probability distribution over possible data outcomes}
\usage{
make_data_probabilities(model, pars, possible_data, A_w = NULL,
  strategy = NULL, strategy_set = NULL, normalize = FALSE)
}
\arguments{
\item{model}{A \code{causal_model}. A model object generated by \code{\link[CausalQueries]{make_model}}}

\item{pars}{A parameter vector}

\item{possible_data}{Possible events data}

\item{A_w}{Ambiguity matrix for data types, optional}

\item{strategy}{A set of node to be sought}

\item{strategy_set}{vector containing possible strategies, optional}

\item{normalize}{logical if TRUE probabilites are normalized to sum to 1}
}
\value{
A dataset
}
\description{
NOTE: This needs to be checked for whether it is taking account of strategy probabilities properly
}
\examples{

model <- make_model("X->M->Y")
possible_data <- make_possible_data(model, N= 2, vars = list(model$node), within = FALSE)
make_data_probabilities(model, pars = get_parameters(model), possible_data)

given <- data.frame(X = c(0,0,0,1,1,1), M = NA, Y = c(0,0,1,0,1,1)) \%>\%
  collapse_data(model, drop_family = TRUE)
possible_data <- make_possible_data(model, observed = given, condition = "X==1 & Y==1", vars = "M", within = TRUE )
make_data_probabilities(model, pars = get_parameters(model), possible_data)

}
