% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy_evaluation.R
\name{strategy_evaluation}
\alias{strategy_evaluation}
\title{Strategy evaluation}
\usage{
strategy_evaluation(model, strategy = c("X", "C", NA), given = "Y==0",
  query, prices = NULL)
}
\description{
A strategy is a triple that indicates (a) first node sought (b) action if first node is 0 (c) action if first node = 1
This function calculates the expected posterior variance from each strategy and the expected number of clues sought
}
\examples{
model <-
  make_model("S -> C -> Y <- R <- X; X -> C -> R") \%>\%
     set_restrictions(
     labels = list(C = c("C1110", "C1111"),
                   R = c("R0001", "R0000"),
                   Y = c("Y0001")), keep = TRUE)
strategies <- two_step_strategies(model)

strategy_evaluation(model,
		strategy  = strategies[1],  given = "Y==0",
		query = "(Y[S=1] != Y[S=0])")

}
