% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy_evaluation.R
\name{strategy_evaluation_single}
\alias{strategy_evaluation_single}
\title{Internal function for evaluating a single strategy}
\usage{
strategy_evaluation_single(model, strategy, query, given = NULL,
  prices = NULL, vars = model$nodes, given_vars = NULL)
}
\description{
Internal function for evaluating a single strategy
}
