% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy_evaluation.R
\name{two_step_strategies}
\alias{two_step_strategies}
\title{Make set of two step strategies}
\usage{
two_step_strategies(model, vars = model$nodes)
}
\arguments{
\item{model}{A \code{causal_model}. A model object generated by \code{\link[gbiqq]{make_model}}}

\item{vars}{Variables to}
}
\description{
Make set of two step strategies
}
\examples{
two_step_strategies(make_model("X->M->Y"))
}
