% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{all_possible}
\alias{all_possible}
\title{Helper for getting all data on specified node with N observed}
\usage{
all_possible(model, N, vars = NULL, condition = TRUE,
  possible_data = TRUE, complete_data = TRUE)
}
\arguments{
\item{model}{model made with gbiqq::make_model}

\item{N}{Integer, number of observed cases}

\item{vars}{String vector listing node observed}

\item{condition}{Statement indicating condition satisfied by observed data}
}
\description{
Helper for getting all data on specified node with N observed
}
\examples{
model <- make_model("X->M->Y")
gbiqqtools:::all_possible(model, N=2, vars = c("X", "M"))
gbiqqtools:::all_possible(model, N=2, vars = c("X", "Y"), condition = "Y==0")
}
