% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data_probabilities.R
\name{average_data_probabilities}
\alias{average_data_probabilities}
\title{Generates an "average" probability distribution over possible data outcomes}
\usage{
average_data_probabilities(model, data_event, using, sims = 500)
}
\arguments{
\item{model}{A causal model as created by \code{make_model}}

\item{data_event}{A data frame with event plus one or multiple data columns}
}
\value{
A dataset
}
\description{
NOTE: This needs to be checked for whether it is taking account of strategy probabilities properly
}
\examples{
library(dplyr)
model <- make_model("X->M->Y")
possible_data <- make_possible_data(model, vars = list(c("X", "M", "Y")), N= 1, within = FALSE)
average_data_probabilities(model, possible_data)


}
