% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data_probabilities.R
\name{make_data_probabilities}
\alias{make_data_probabilities}
\title{Generates a probability distribution over possible data outcomes}
\usage{
make_data_probabilities(model, pars, possible_data, A_w = NULL,
  strategy = NULL, strategy_set = NULL)
}
\arguments{
\item{model}{A causal model as created by \code{make_model}}

\item{pars}{A parameter vector}

\item{possible_data}{Possible events data}

\item{A_w}{Ambiguity matrix for data types, optional}

\item{strategy}{vector providing data strategy set for event, optional}

\item{strategy_set}{vector containing possible strategies, optional}
}
\value{
A dataset
}
\description{
NOTE: This needs to be checked for whether it is taking account of strategy probabilities properly
}
\examples{

library(dplyr)
model <- make_model("X->M->Y") \%>\%
         set_parameters(type = "flat")
possible_data <- make_possible_data(model, N= 2, vars = list(model$variables), within = FALSE)
make_data_probabilities(model, pars = model$parameters, possible_data)

given <- data.frame(X = c(0,0,0,1,1,1), M = NA, Y = c(0,0,1,0,1,1)) \%>\%
  collapse_data(model, remove_family = TRUE)
possible_data <- make_possible_data(model, given = given, condition = "X==1 & Y==1", vars = "M", within = TRUE )
make_data_probabilities(model, pars = model$parameters, possible_data)

}
