% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_estimates_database.R
\name{make_estimates_database}
\alias{make_estimates_database}
\title{Generates a database of results using gbiqq over possible data}
\usage{
make_estimates_database(
  model,
  observed,
  possible_data = NULL,
  queries = "Y[X=1]>Y[X=0]",
  given = TRUE,
  expand_grid = FALSE,
  use_parameters = FALSE,
  iter = 4000,
  refresh = 0,
  chains = NULL,
  ...
)
}
\arguments{
\item{model}{A \code{causal_model}. A model object generated by \code{\link[gbiqq]{make_model}}}

\item{observed}{A data.frame with observations.}

\item{possible_data}{A data frame with an events column and possible data columns (if a strategy columns is included it is ignored)}

\item{queries}{list of statements for causal queries}

\item{given}{A conditioning set as a character string that evaluates to a logical, for example 'Y==1'}

\item{expand_grid}{logical, If TRUE combinations of queries and given are expanded}

\item{iter}{Integer, passed to stan, defaults to 4000}

\item{chains}{Integer, passed to stan, defaults to 4}
}
\value{
A list with query output dataframes for each data strategy
}
\description{
This function can run many models and can take a long time depending on the size of possible data.
}
\examples{
model <- make_model("X->M->Y")  \%>\%
   set_restrictions(c("Y[M=1]<Y[M=0]", "M[X=1]<M[X=0]")) \%>\%
   set_parameter_matrix()

observed <- data.frame(X = c(0,0,0,1,1,1), M = NA, Y = c(0,0,1,0,1,1)) \%>\%
        collapse_data(model, drop_family = TRUE)


possible_data <- make_possible_data(model, observed, vars = "M", condition = "X==1 & Y==1")

estimates_database <- make_estimates_database(
      model,
      observed = observed,
      possible_data = possible_data,
      queries = "Y[X=1]>Y[X=0]")

estimates_database <- make_estimates_database(
      model,
      observed = observed,
      possible_data = possible_data,
      queries = c(ATE = "Y[X=1]-Y[X=0]", PC = "Y[X=1]>Y[X=0]"),
      given = c(TRUE, "Y==1 & X==1"))
}
