% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy_evaluation.R
\name{strategy_evaluation_single}
\alias{strategy_evaluation_single}
\title{Internal function for evaluating a single strategy}
\usage{
strategy_evaluation_single(model, strategy, query, given = NULL,
  prices = NULL, vars = model$nodes, given_vars = NULL)
}
\arguments{
\item{model}{A \code{causal_model}. A model object generated by \code{\link[CausalQueries]{make_model}}}

\item{strategy}{A set of node to be sought}

\item{query}{A character string. An expression defining nodal types to interrogate \code{\link[CausalQueries]{reveal_outcomes}}}

\item{given}{A conditioning set as a character string that evaluates to a logical, for example 'Y==1'}
}
\description{
Internal function for evaluating a single strategy
}
\keyword{internal}
