% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy_evaluation.R
\name{two_step_strategies}
\alias{two_step_strategies}
\title{Make set of two step strategies}
\usage{
two_step_strategies(model, vars = model$nodes)
}
\description{
Make set of two step strategies
}
